Move 'Todo' model into 'todo_lists' app
